requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning
xgboost>=1.5.0
lightgbm>=3.2.0
catboost>=1.0.0

# Explainability
shap>=0.40.0

# Configuration
pyyaml>=6.0
omegaconf>=2.1.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
ruff>=0.0.200
pre-commit>=2.15.0

# Optional: Deep learning (uncomment if needed)
# torch>=1.10.0
# pytorch-lightning>=1.5.0

# Optional: Web framework for demo
streamlit>=1.0.0
gradio>=3.0.0

# Optional: Additional ML libraries
# imbalanced-learn>=0.8.0
# optuna>=2.10.0
